requirements.txt
PySide6>=6.9.0
opencv-python>=4.8.0
numpy>=1.24.0
websockets>=12.0
paho-mqtt>=1.6.1
